do not delete Chrome downloaded files on Linux
https://github.com/SeleniumHQ/selenium/issues/5292

git-svn-id: http://svn.redmine.org/redmine/trunk@18531 e93f8b46-1217-0410-a6f0-8f06a7374b81

+++ test/application_system_test_case.rb
@@ -59,7 +59,8 @@ class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
   end
 
   def clear_downloaded_files
-    FileUtils.rm downloaded_files
+    # https://github.com/SeleniumHQ/selenium/issues/5292
+    FileUtils.rm downloaded_files if Redmine::Platform.mswin?
   end
 
   def downloaded_files(filename='*')
